chore: rename guides to howto
